// src/types/inquiriesType.ts
export type Json = string | number | boolean | null | { [key: string]: Json | undefined } | Json[];

export type inquirie = Database['public']['Tables']['user_inquiries']['Row'];
export type inquirieInsert = Database['public']['Tables']['user_inquiries']['Insert'];
export type inquirieUpdate = Database['public']['Tables']['user_inquiries']['Update'];

export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: '13.0.5';
  };
  public: {
    Tables: {
// ...
      user_inquiries: {
        Row: {
          inquiry_answer: string | null;
          inquiry_answered_at: string | null;
          inquiry_created_at: string;
          inquiry_detail: string;
          inquiry_file_urls: Json | null;
          inquiry_id: string;
          inquiry_main_type: string;
          inquiry_status: Database['public']['Enums']['inquiry_status_enum'];
          inquiry_sub_type: string;
          user_id: string;
        };
        Insert: {
          inquiry_answer?: string | null;
          inquiry_answered_at?: string | null;
          inquiry_created_at?: string;
          inquiry_detail: string;
          inquiry_file_urls?: Json | null;
          inquiry_id?: string;
          inquiry_main_type: string;
          inquiry_status?: Database['public']['Enums']['inquiry_status_enum'];
          inquiry_sub_type: string;
          user_id: string;
        };
        Update: {
          inquiry_answer?: string | null;
          inquiry_answered_at?: string | null;
          inquiry_created_at?: string;
          inquiry_detail?: string;
          inquiry_file_urls?: Json | null;
          inquiry_id?: string;
          inquiry_main_type?: string;
          inquiry_status?: Database['public']['Enums']['inquiry_status_enum'];
          inquiry_sub_type?: string;
          user_id?: string;
        };
        Relationships: [
          {
            foreignKeyName: 'user_inquiries_user_id_fkey';
            columns: ['user_id'];
            isOneToOne: false;
            referencedRelation: 'user_profiles';
            referencedColumns: ['user_id'];
          },
        ];
      };
// ...
    };
// ...
  };
};
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema['Enums']
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions['schema']]['Enums']
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions['schema']]['Enums'][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema['Enums']
    ? DefaultSchema['Enums'][DefaultSchemaEnumNameOrOptions]
    : never;